webapp/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packageurl-python==0.11.1
packaging==23.1
pip-requirements-parser==32.0.1
psycopg2-binary==2.9.6
pyparsing==3.1.0
python-dotenv==1.0.0
sortedcontainers==2.4.0
toml==0.10.2
Werkzeug==2.3.6
